prov/efa: move protocol v4 definition to a separate header file
This patch moves all constants, flags, and data structures in
protocol v4 to a separate header file rdm_proto_v4.h.

This is to separate protocol definition and implementation.
